+++ h2integrate/finances/profast_finance.py
@@ -436,13 +436,29 @@ class ProFastComp(om.ExplicitComponent):
         self.options.declare("commodity_type", types=str)
         self.options.declare("description", types=str, default="")
 
+    def add_model_specific_outputs(self):
+        self.add_output(self.LCO_str, val=0.0, units=self.lco_units)
+        self.outputs_to_units = {
+            "wacc": "percent",
+            "crf": "percent",
+            "irr": "percent",
+            "profit_index": "unitless",
+            "investor_payback_period": "yr",
+            "price": self.lco_units,
+        }
+        for output_var, units in self.outputs_to_units.items():
+            self.add_output(f"{output_var}_{self.output_txt}", val=0.0, units=units)
+
+        self.add_discrete_output(f"{self.LCO_str}_breakdown", val={}, desc="LCO Breakdown of costs")
+        return
+
     def setup(self):
         if self.options["commodity_type"] == "electricity":
             commodity_units = "kW*h/year"
-            lco_units = "USD/kW/h"
+            self.lco_units = "USD/kW/h"
         else:
             commodity_units = "kg/year"
-            lco_units = "USD/kg"
+            self.lco_units = "USD/kg"
 
         LCO_base_str = f"LCO{self.options['commodity_type'][0].upper()}"
         self.output_txt = self.options["commodity_type"].lower()
@@ -456,17 +472,7 @@ class ProFastComp(om.ExplicitComponent):
                 self.output_txt = f"{self.options['commodity_type'].lower()}_{desc_str}"
                 self.LCO_str = f"{LCO_base_str}_{desc_str}"
 
-        self.add_output(self.LCO_str, val=0.0, units=lco_units)
-        self.outputs_to_units = {
-            "wacc": "percent",
-            "crf": "percent",
-            "irr": "percent",
-            "profit_index": "unitless",
-            "investor_payback_period": "yr",
-            "price": lco_units,
-        }
-        for output_var, units in self.outputs_to_units.items():
-            self.add_output(f"{output_var}_{self.output_txt}", val=0.0, units=units)
+        self.add_model_specific_outputs()
 
         if self.options["commodity_type"] == "co2":
             self.add_input("co2_capture_kgpy", val=0.0, units="kg/year")
@@ -514,7 +520,7 @@ class ProFastComp(om.ExplicitComponent):
             "variable_costs", {}
         )
         variable_cost_params.setdefault("escalation", self.params.inflation_rate)
-        variable_cost_params.setdefault("unit", lco_units.replace("USD", "$"))
+        variable_cost_params.setdefault("unit", self.lco_units.replace("USD", "$"))
         self.variable_cost_settings = ProFASTDefaultVariableCost.from_dict(variable_cost_params)
 
         # incentives - unused for now
@@ -524,7 +530,7 @@ class ProFastComp(om.ExplicitComponent):
         # incentive_params.setdefault("decay", -1 * self.params.inflation_rate)
         # self.incentive_params_settings = ProFASTDefaultIncentive.from_dict(incentive_params)
 
-    def compute(self, inputs, outputs):
+    def populate_profast(self, inputs):
         mass_commodities = ["hydrogen", "ammonia", "co2", "nitrogen", "methanol"]
 
         years_of_operation = create_years_of_operation(
@@ -639,10 +645,29 @@ class ProFastComp(om.ExplicitComponent):
         pf_dict["fixed_costs"] = fixed_costs
         pf_dict["feedstocks"] = variable_costs
         # create ProFAST object
-
         pf = create_and_populate_profast(pf_dict)
+        return pf
+
+    def compute(self, inputs, outputs, discrete_inputs, discrete_outputs):
+        pf = self.populate_profast(inputs)
+
         # simulate ProFAST
         sol, summary, price_breakdown = run_profast(pf)
+
+        outputs[self.LCO_str] = sol["lco"]
+        for output_var in self.outputs_to_units.keys():
+            val = sol[output_var.replace("_", " ")]
+            if isinstance(val, (np.ndarray, list, tuple)):  # only for IRR
+                # if len(val)>0:
+                val = val[-1]
+            outputs[f"{output_var}_{self.output_txt}"] = val
+
+        # make dictionary of ProFAST config
+        pf_config_dict = convert_pf_to_dict(pf)
+
+        # make LCO cost breakdown
+        lco_breakdown, lco_check = make_price_breakdown(price_breakdown, pf_config_dict)
+        discrete_outputs[f"{self.LCO_str}_breakdown"] = lco_breakdown
 
         # Check whether to export profast object to .yaml file
         save_results = self.options["plant_config"]["finance_parameters"]["model_inputs"].get(
@@ -651,9 +676,8 @@ class ProFastComp(om.ExplicitComponent):
         save_config = self.options["plant_config"]["finance_parameters"]["model_inputs"].get(
             "save_profast_config", False
         )
-        if save_results or save_config:
-            pf_config_dict = convert_pf_to_dict(pf)
 
+        if save_results or save_config:
             output_dir = self.options["driver_config"]["general"]["folder_output"]
             fdesc = self.options["plant_config"]["finance_parameters"]["model_inputs"].get(
                 "profast_output_description", "ProFastComp"
@@ -663,22 +687,17 @@ class ProFastComp(om.ExplicitComponent):
 
             output_dir = Path(output_dir)
             output_dir.mkdir(parents=True, exist_ok=True)
+
+            pf_config_dict = dict_to_yaml_formatting(pf_config_dict)
+
             if save_config:
-                pf_config_dict = dict_to_yaml_formatting(pf_config_dict)
                 config_fpath = Path(output_dir) / f"{fbasename}_config.yaml"
                 write_yaml(pf_config_dict, config_fpath)
+
             if save_results:
-                lco_breakdown, lco_check = make_price_breakdown(price_breakdown, pf_config_dict)
                 price_breakdown_formatted = format_profast_price_breakdown_per_year(price_breakdown)
                 pf_breakdown_fpath = Path(output_dir) / f"{fbasename}_profast_price_breakdown.csv"
                 lco_breakdown_fpath = Path(output_dir) / f"{fbasename}_LCO_breakdown.yaml"
                 price_breakdown_formatted.to_csv(pf_breakdown_fpath)
                 lco_breakdown = dict_to_yaml_formatting(lco_breakdown)
                 write_yaml(lco_breakdown, lco_breakdown_fpath)
-
-        outputs[self.LCO_str] = sol["lco"]
-        for output_var in self.outputs_to_units.keys():
-            val = sol[output_var.replace("_", " ")]
-            if isinstance(val, (np.ndarray, list, tuple)):  # only for IRR
-                val = val[-1]
-            outputs[f"{output_var}_{self.output_txt}"] = val